datavex-main/info_gather/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
lxml>=5.1.0
httpx==0.25.2
pdfminer.six==20221105
feedparser==6.0.10
PyGithub==2.1.1
sqlalchemy==2.0.23
psycopg[binary]>=3.1.0
python-dotenv==1.0.0
tenacity==8.2.3
loguru==0.7.2
tqdm==4.66.1
